# Flatten ElementPathTree.