--- a/editor/src/components/editor/store/store-deep-equality-instances.ts
+++ b/editor/src/components/editor/store/store-deep-equality-instances.ts
@@ -503,7 +503,7 @@ import {
 import {
   elementPathTree,
   ElementPathTree,
-  ElementPathTreeRoot,
+  ElementPathTrees,
 } from '../../../core/shared/element-path-tree'
 
 export function TransientCanvasStateFilesStateKeepDeepEquality(
@@ -2143,7 +2143,7 @@ export const CanvasControlTypeKeepDeepEquality: KeepDeepEqualityCall<CanvasContr
   return keepDeepEqualityResult(newValue, false)
 }
 
-export function ElementPathTreeRootKeepDeepEquality(): KeepDeepEqualityCall<ElementPathTreeRoot> {
+export function ElementPathTreesKeepDeepEquality(): KeepDeepEqualityCall<ElementPathTrees> {
   return objectDeepEquality(ElementPathTreeKeepDeepEquality)
 }
 
@@ -2183,7 +2183,7 @@ export const InteractionSessionKeepDeepEquality: KeepDeepEqualityCall<Interactio
     (session) => session.aspectRatioLock,
     nullableDeepEquality(createCallWithTripleEquals()),
     (session) => session.latestElementPathTree,
-    ElementPathTreeRootKeepDeepEquality(),
+    ElementPathTreesKeepDeepEquality(),
     interactionSession,
   )
 
@@ -3775,7 +3775,7 @@ export const EditorStateKeepDeepEquality: KeepDeepEqualityCall<EditorState> = (
     oldValue.jsxMetadata,
     newValue.jsxMetadata,
   )
-  const elementPathTreeResult = ElementPathTreeRootKeepDeepEquality()(
+  const elementPathTreeResult = ElementPathTreesKeepDeepEquality()(
     oldValue.elementPathTree,
     newValue.elementPathTree,
   )