… editor/src/components/inspector/inspector-strategies/inspector-strategy.ts
… editor/src/components/navigator/navigator-item/navigator-item-dnd-container.tsx
… editor/src/components/navigator/navigator-utils.ts
… editor/src/core/model/conditionals.ts
… editor/src/core/model/element-metadata-utils.ts
… editor/src/core/shared/element-path-tree.benchmark.ts
… editor/src/core/shared/element-path-tree.ts