**Context:**
It was observed that if `MetadataUtils.getChildrenOrdered` could be made pretty inexpensive to call, then the alternative unordered variants of functions could be eliminated and then functions like `getElementWarningsInner` could be made to perform significantly faster (in our example project that can breach 10ms on its own). As such, while this does provide some performance benefit of its own, it's primarily a stepping stone to be able to make those changes.

**Problem:**
`ElementPathTreeRoot` currently requires a tree traversal to find an entry so that the ordered children can be identified for the target element, which gets increasingly expensive the deeper the tree is.

**Fix:**
`ElementPathTreeRoot` is now flattened out to contain every single `ElementPathTree` indexed by their path, which completely eliminates the need for traversals and means that now getting a sub tree is an O(1) operation. Where it used to be O(n) where n is the depth of the target item in the tree.

**Benchmarks:**
```
buildTree - deeply nested elements:
481 ops/s -> 464 ops/s
buildTree - very wide elements:
44,817 ops/s -> 41,609 ops/s
getSubTree - deeply nested elements:
49,871 ops/s -> 68,355,704 ops/s
getSubTree - very wide elements:
5,882,326 ops/s -> 71,411,579 ops/s
```

Skimming a performance profile in Chrome, `MetadataUtils.getChildrenOrdered` is now small enough that it doesn't even show up at all.

**Commit Details:**
- Renamed `ElementPathTreeRoot` to `ElementPathTrees`.
- Changed `ElementPathTree.children` to be an array of `ElementPathTree`.
- Added `ElementPathTree.pathString` for convenience.
- Refactored `buildTree` to cater for the change to `ElementPathTree`.
- Added `getStoryboardRoot` and `getCanvasRoots` utility functions.
- Reworked `reorderTree` to handle the `ElementPathTree` changes and to now mutate the tree.
- `printTree` now uses `getStoryboardRoot`.
- `getSubTree` now rewritten to be significantly simpler.
- `createOrderedElementPathsFromElements` now uses `getCanvasRoots` defined elsewhere.
- `createElementPathTreeFromMetadata` refactored due to `ElementPathTree` changes.
- Small change to `reorderConditionalChildPathTrees`.
- `getNavigatorTargets` now uses `getCanvasRoots` defined elsewhere.
- Added benchmarks for `getSubTree`.